requirements.txt
aversine==0.0.1
pytest==7.1.3